Testing the morale system.
Tested the morale API. Created a CLI testing method using the app.js entry point.
In order to stop the clock and end the program the clock.pause() needs to be called at a certain point. When calling
clock.pause(), keep in mind that the pause method emits an update event, and this might incur a circular update event calls.

Next task is to test and refactor the resources feature. And inlcude each resource gathering time cost.